requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jieba==0.42.1
httpx==0.27.2
pydantic==2.9.2
python-dotenv==1.0.1
tenacity==9.0.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0

